feat:refactor files from types group to feature group closes #21
