--- constants/menu_id.js
@@ -1,6 +1,0 @@
-export const MENU_IDS = {
-  focus: 'start-focusing',
-  shortBreak: 'start-short-break',
-  longBreak: 'start-long-break'
-};
-
