--- constants/mode.js
@@ -1,6 +1,0 @@
-export const MODES = {
-    focus: 'focus',
-    longBreak: 'longBreak',
-    shortBreak: 'shortBreak'
-}
-
