--- constants/status.js
@@ -1,6 +1,0 @@
-export const STATUSES = {
-  idle: 'idle',
-  running: 'running',
-  paused: 'paused'
-};
-
