+++ src/core/constants.js
@@ -1,0 +1,18 @@
+export const MENU_IDS = {
+  focus: 'start-focusing',
+  shortBreak: 'start-short-break',
+  longBreak: 'start-long-break'
+};
+
+export const MODES = {
+    focus: 'focus',
+    longBreak: 'longBreak',
+    shortBreak: 'shortBreak'
+}
+
+export const STATUSES = {
+  idle: 'idle',
+  running: 'running',
+  paused: 'paused'
+};
+
