+++ src/features/messaging/messing_handler.js
@@ -1,5 +1,5 @@
-import { startPomodoroAlarm } from "../services/alarm_service.js";
-import { executeWithManager } from "../services/pomodoro_service.js";
+import { startPomodoroAlarm } from "../alarm/alarm_handler.js";
+import { executeWithManager } from "../../core/timer_executor.js";
 
 export async function onStartSessionMessage(mode) {
   await executeWithManager(async manager => {
